+++ mtgjson5/providers/scryfall/set_language_detector.py
@@ -1,4 +1,3 @@
-import json
 import logging
 import time
 from typing import Any, Dict, List, Optional, Union
